+++ JeMPI_Apps/JeMPI_UI/src/pages/settings/deterministic/SourceView.tsx
@@ -1,0 +1,144 @@
+import React, { useState, useEffect } from 'react';
+import Box from '@mui/material/Box';
+import { DataGrid, GridColDef, GridEventListener, GridRowEditStopReasons, GridRowId, GridRowModel, GridRowModes, GridRowModesModel, GridActionsCellItem } from '@mui/x-data-grid';
+import EditIcon from '@mui/icons-material/Edit';
+import SaveIcon from '@mui/icons-material/Save';
+import CancelIcon from '@mui/icons-material/Close';
+
+interface RowData {
+  id: string;
+  ruleNumber: number;
+  ruleText: string;
+}
+
+interface SourceViewProps {
+  data: RowData[];
+}
+
+const SourceView: React.FC<SourceViewProps> = ({ data }) => {
+  const [rows, setRows] = useState<RowData[]>([]);
+  const [rowModesModel, setRowModesModel] = useState<GridRowModesModel>({});
+
+  useEffect(() => {
+    setRows(data);
+  }, [data]);
+
+  const handleEditClick = (id: GridRowId) => () => {
+    setRowModesModel({ ...rowModesModel, [id]: { mode: GridRowModes.Edit } });
+  };
+
+  const handleSaveClick = (id: GridRowId) => () => {
+    setRowModesModel({ ...rowModesModel, [id]: { mode: GridRowModes.View } });
+  };
+
+  const handleCancelClick = (id: GridRowId) => () => {
+    setRowModesModel({
+      ...rowModesModel,
+      [id]: { mode: GridRowModes.View, ignoreModifications: true }
+    });
+  };
+
+  const processRowUpdate = (newRow: GridRowModel) => {
+    const { id, ...updatedRow } = newRow;
+    setRows(rows.map(row => (row.id === id ? updatedRow as RowData : row)));
+    return updatedRow as RowData;
+  };
+
+  const handleRowModesModelChange = (newRowModesModel: GridRowModesModel) => {
+    setRowModesModel(newRowModesModel);
+  };
+
+  const handleRowEditStop: GridEventListener<'rowEditStop'> = (params, event) => {
+    if (params.reason === GridRowEditStopReasons.rowFocusOut) {
+      event.defaultMuiPrevented = true;
+    }
+  };
+
+  const columns: GridColDef[] = [
+    {
+      field: 'ruleNumber',
+      headerName: 'Rule Number',
+      width: 150,
+      align: 'center',
+      headerAlign: 'center',
+      editable: false,
+    },
+    {
+      field: 'ruleText',
+      headerName: 'Rule',
+      width: 500,
+      align: 'left',
+      headerAlign: 'left',
+      editable: true,
+    },
+    {
+      field: 'actions',
+      type: 'actions',
+      headerName: 'Actions',
+      width: 150,
+      align: 'center',
+      headerAlign: 'center',
+      cellClassName: 'actions',
+      getActions: ({ id }) => {
+        const isInEditMode = rowModesModel[id]?.mode === GridRowModes.Edit;
+        if (isInEditMode) {
+          return [
+            <GridActionsCellItem
+              icon={<SaveIcon />}
+              id="save-button"
+              label="Save"
+              sx={{ color: 'white' }}
+              onClick={handleSaveClick(id)}
+            />,
+            <GridActionsCellItem
+              icon={<CancelIcon />}
+              id="cancel-button"
+              label="Cancel"
+              className="textPrimary"
+              onClick={handleCancelClick(id)}
+              color="inherit"
+            />
+          ];
+        }
+
+        return [
+          <GridActionsCellItem
+            icon={<EditIcon />}
+            id="edit-button"
+            label="Edit"
+            className="textPrimary"
+            onClick={handleEditClick(id)}
+            color="inherit"
+          />
+        ];
+      }
+    }
+  ];
+
+  return (
+    <Box
+      sx={{
+        height: 500,
+        width: '100%',
+        '& .actions': {
+          color: '#fff'
+        },
+        '& .textPrimary': {
+          color: '#fff'
+        }
+      }}
+    >
+      <DataGrid
+        rows={rows}
+        columns={columns}
+        editMode="row"
+        rowModesModel={rowModesModel}
+        onRowModesModelChange={handleRowModesModelChange}
+        onRowEditStop={handleRowEditStop}
+        processRowUpdate={processRowUpdate}
+      />
+    </Box>
+  );
+};
+
+export default SourceView;
